+++ docs/CHANGELOG.md
@@ -5,6 +5,11 @@
  The format is based on [Keep a Changelog](https://keepachangelog.com/en/1/0/0/),
  and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+- Fix QR ability for Endroid v6 by using constructor API instead of deprecated QrCode::create().
+
 ## [2.1.1] - 2026-03-04
 
 ### Fixed